+++ src/main/java/kz/kakimzhanova/delivery/service/OrderService.java
@@ -3,6 +3,7 @@ package kz.kakimzhanova.delivery.service;
 import kz.kakimzhanova.delivery.entity.Order;
 import kz.kakimzhanova.delivery.exception.ServiceException;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface OrderService {
@@ -10,5 +11,5 @@ public interface OrderService {
     Order newOrder(String login) throws ServiceException;
     boolean deleteOrder(int orderId) throws ServiceException;
     Order findOrderById(int orderId) throws ServiceException;
-    boolean updateTotalCost(int orderId, float totalCost) throws ServiceException;
+    boolean updateTotalCost(int orderId, BigDecimal totalCost) throws ServiceException;
 }
